requirements.txt
selenium
openai
python-dotenv
undetected-chromedriver
docx2txt
pdfplumber